Added ability to delete

+++ LearningDiaryJS/Models/Topic.cs
@@ -10,8 +10,8 @@ namespace LearningDiaryJS.Models
         public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public int? TimeToMaster { get; set; }
-        public int? TimeSpent { get; set; }
+        public double? TimeToMaster { get; set; }
+        public double? TimeSpent { get; set; }
         public string Source { get; set; }
         public DateTime? StartLearningDate { get; set; }
         public bool? InProgress { get; set; }
